refactors StudentRecordService to ApiService
